requirements.txt
astrapy>=2.0.0,<3.0.0
fastapi>=0.115.0
langchain>=0.3
langchain_community>=0.3
langchain_openai>=0.3
python-dotenv>=1.0.0
tiktoken>=0.9.0
uvicorn[standard]>=0.34.0
